fix: 폴더 구조 변경 / step state로 컴포넌트 변경
