+++ src/app/record/(info)/common/common.css.ts
@@ -2,7 +2,6 @@ import { style } from "@vanilla-extract/css";
 
 export const infoWrapper = style({
   height: "100%",
-  padding: "85px 36px 104px 36px",
 });
 
 export const infoContainer = style({
