+++ src/app/record/(info)/detail/components/popup.tsx
