+++ src/app/record/(info)/detail/components/titleText.tsx
