+++ src/app/record/(info)/detail/page.tsx
@@ -5,13 +5,14 @@ import Memo from "@/app/components/common/Memo";
 import { flexSprinklesFc } from "@/app/components/common/utils/flex";
 import { colors } from "@/app/styles/colors.css";
 
-import { infoContainer } from "../common/common.css";
 import useInfoStore from "@/app/store/info/infoStore";
 import { usePopupStore } from "@/app/store/popup/PopupStore";
 import DetailPopup from "./components/popup";
 import TitleText from "./components/titleText";
+import { infoContainer } from "../common/common.css";
+import { StepChangeHandler } from "../../page";
 
-const Page = () => {
+const DetailPage = ({ onButtonClick }: { onButtonClick: StepChangeHandler }) => {
   const setRecordNoteState = useInfoStore((state) => state.setRecordNote);
   const setDetailPopupState = usePopupStore((state) => state.setIsPopup);
   const saveRecord = useInfoStore((state) => state.saveRecord);
@@ -23,6 +24,7 @@ const Page = () => {
     } else {
       setDetailPopupState(true);
       saveRecord();
+      onButtonClick(0);
     }
   };
   return (
@@ -49,4 +51,4 @@ const Page = () => {
   );
 };
 
-export default Page;
+export default DetailPage;
