+++ src/app/record/(info)/shape/components/consistencyBox.tsx
