+++ src/app/record/(info)/shape/components/contentBox.tsx
