+++ src/app/record/(info)/shape/components/shapeButton.tsx
@@ -1,22 +1,9 @@
 import Button from "@/app/components/common/Button";
 import { flexSprinklesFc } from "@/app/components/common/utils/flex";
-import useInfoStore from "@/app/store/info/infoStore";
-
+import { StepChangeHandler } from "@/app/record/page";
 import { colors } from "@/app/styles/colors.css";
-import { useRouter } from "next/navigation";
 
-const TimeButton = () => {
-  const router = useRouter();
-  const bowelTime = useInfoStore((state) => state.bowelTime);
-
-  const onNext = () => {
-    if (bowelTime.hour === 0 && bowelTime.minute === 0) {
-      alert("시간을 선택하세요");
-      return;
-    }
-    router.push("/info/shape");
-  };
-
+const ShapeButton = ({ onButtonClick }: { onButtonClick: StepChangeHandler }) => {
   return (
     <div className={flexSprinklesFc({ gap: "16px", justifyContent: "center" })}>
       <Button
@@ -24,7 +11,7 @@ const TimeButton = () => {
         height="59px"
         borderRadius="10px"
         onClick={() => {
-          router.push("/record");
+          onButtonClick(1);
         }}
       />
       <Button
@@ -34,10 +21,12 @@ const TimeButton = () => {
         background={colors.primary}
         color={colors.white}
         borderRadius="10px"
-        onClick={onNext}
+        onClick={() => {
+          onButtonClick(3);
+        }}
       />
     </div>
   );
 };
 
-export default TimeButton;
+export default ShapeButton;
