+++ src/app/record/(info)/shape/components/shapeTypeBox.tsx
