+++ src/app/record/(info)/shape/styles/shape.css.ts
