+++ src/app/record/(info)/time/page.tsx
@@ -1,11 +1,12 @@
 "use client";
 
+import { StepChangeHandler } from "../../page";
 import { infoContainer } from "../common/common.css";
 import TimeButton from "./components/timeButton";
 import TimePicker from "./components/timePicker";
 import TimeTitle from "./components/timeTitle";
 
-const Page = () => {
+const TimePage = ({ onButtonClick }: { onButtonClick: StepChangeHandler }) => {
   return (
     <article className={infoContainer}>
       <TimeTitle />
@@ -13,9 +14,9 @@ const Page = () => {
         <TimePicker />
       </div>
 
-      <TimeButton />
+      <TimeButton onButtonClick={onButtonClick} />
     </article>
   );
 };
 
-export default Page;
+export default TimePage;
